Move sync script
